pca945x_select fixes, missing args etc

+++ artiq/firmware/libproto_artiq/kernel_proto.rs
@@ -65,7 +65,7 @@ pub enum Message<'a> {
     I2cReadRequest { busno: u32, ack: bool },
     I2cReadReply { succeeded: bool, data: u8 },
     I2cBasicReply { succeeded: bool },
-    I2cPca954xSelectRequest { busno: u32, address: u32, channel: u32, clear: bool },
+    I2cPca954xSelectRequest { busno: u32, address: u8, channel: u8, clear: bool },
 
     SpiSetConfigRequest { busno: u32, flags: u8, length: u8, div: u8, cs: u8 },
     SpiWriteRequest { busno: u32, data: u32 },
